feat: created a dedicated folder for dtos
