--- src/main/java/com/example/redditvault/client/DownloadRequest.java
@@ -1,29 +1,0 @@
-package com.example.redditvault.client;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class DownloadRequest {
-    private String filename;
-    private String url;
-
-    public DownloadRequest(String url, String filename) {
-        this.url = url;
-        this.filename = filename;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getFilename() {
-        return filename;
-    }
-
-    public void setFilename(String filename) {
-        this.filename = filename;
-    }
-}
