+++ src/main/java/com/example/redditvault/client/MediaExtractorRegistry.java
@@ -1,5 +1,7 @@
 package com.example.redditvault.client;
 
+import com.example.redditvault.client.dto.RedditSavedItem;
+
 import java.util.List;
 
 public class MediaExtractorRegistry {
