+++ src/main/java/com/example/redditvault/client/dto/MediaMetadata.java
@@ -1,4 +1,4 @@
-package com.example.redditvault.client;
+package com.example.redditvault.client.dto;
 
 public class MediaMetadata {
     private MediaSource source;
