+++ src/main/java/com/example/redditvault/client/dto/MediaSource.java
@@ -1,4 +1,4 @@
-package com.example.redditvault.client;
+package com.example.redditvault.client.dto;
 
 public class MediaSource {
     private String url;
