+++ src/main/java/com/example/redditvault/client/dto/RedditProperties.java
@@ -1,4 +1,4 @@
-package com.example.redditvault.client;
+package com.example.redditvault.client.dto;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.http.*;
