+++ src/main/java/com/example/redditvault/client/dto/RedditVideo.java
@@ -1,4 +1,4 @@
-package com.example.redditvault.client;
+package com.example.redditvault.client.dto;
 
 public class RedditVideo {
     //TODO: add field for audio URL
